requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
requests!=2.9.0,>=2.8.1 # Apache-2.0
PyYAML>=3.10.0 # MIT
boto3>=1.4.4
